add compatibility for mobile phone
