+++ app/styles/main.css
@@ -1546,6 +1546,47 @@ div div.contaxt-menu-rephrase button.rephrase-btn:hover{
   height: 50px;
   margin: 0 40px;
 }
+.logos1 {
+  overflow: hidden;
+  padding: 30px 0;
+  background: white;
+  white-space: nowrap;
+  position: relative;
+  width: 100%;
+}
+
+.logos1:before,
+.logos1:after {
+  position: absolute;
+  top: 0;
+  width: 50px;
+  height: 100%;
+  content: "";
+  z-index: 2;
+}
+
+.logos1:before {
+  left: 0;
+  background: linear-gradient(to left, rgba(255, 255, 255, 0), white);
+}
+
+.logos1:after {
+  right: 0;
+  background: linear-gradient(to right, rgba(255, 255, 255, 0), white);
+}
+
+.logos1:hover .logos-slide1 {
+  animation-play-state: paused;
+}
+
+.logos-slide1 {
+    animation: 20s slide infinite linear;
+}
+
+.logos-slide1 img {
+  height: 50px;
+  margin: 0 10px;
+}
 /* _____________________account css ____________________*/
 .container-login{
   background-color: #FAFAFA;
@@ -1799,6 +1840,66 @@ div div.contaxt-menu-rephrase button.rephrase-btn:hover{
 }
 
 .article-details p {
+  margin: 0;
+  color: #666;
+}
+/* ____ */
+.articles-container1 {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+.article-card1 {
+  display: flex;
+  align-items: flex-start;
+  border: 1px solid #ddd;
+  padding: 15px;
+  margin: 10px 0;
+  width:100%;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  transition: box-shadow 0.3s ease-in-out;
+  text-align: start;
+}
+
+.article-card1:hover {
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+}
+
+.article-cover1 {
+  margin-right: 20px;
+}
+
+.article-cover1 img {
+  width: 50px;
+  height: 50px;
+  border-radius: 5px;
+}
+
+.article-details1{
+  display: flex;
+  flex-direction: column;
+}
+
+.article-meta1 {
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+  color: #666;
+}
+
+.article-meta1 span {
+  margin-right: 10px;
+  color: #007bff; /* blue color for type */
+}
+
+.article-details1 h2 {
+  margin: 0;
+  font-size: 1.2em;
+}
+
+.article-details1 p {
   margin: 0;
   color: #666;
 }
